feat(FR-2494): extract Legacy auto-scaling rule components and add version gating (#6639)
Resolves #6485(FR-2494) — Sub-task 1: Legacy extraction + version gating

## Summary

- Rename `AutoScalingRuleEditorModal` to `AutoScalingRuleEditorModalLegacy` with updated GraphQL operation names
- Extract the auto-scaling rules table from `EndpointDetailPage` into a new `AutoScalingRuleListLegacy` component
- Add `prometheus-auto-scaling-rule` feature key gated to `>=26.4.0` for future Strawberry API path
- Apply comparator display normalization in Legacy list: `GREATER_THAN` is flipped to show `[threshold] < [metric_name]`
- Limit Legacy editor comparator dropdown to only `LESS_THAN` and `GREATER_THAN` (removed `≤` and `≥` options)
- Set up version gating conditional in `EndpointDetailPage` with placeholder for Strawberry path (`>=26.4.0`)

## Changed files

- `react/src/components/AutoScalingRuleEditorModal.tsx` → renamed to `AutoScalingRuleEditorModalLegacy.tsx`
- `react/src/components/AutoScalingRuleListLegacy.tsx` (new)
- `react/src/pages/EndpointDetailPage.tsx` (simplified, uses extracted components)
- `src/lib/backend.ai-client-esm.ts` (added `prometheus-auto-scaling-rule` key)

## Verification

```
=== Relay: PASS ===
=== Lint: PASS ===
=== Format: PASS ===
=== TypeScript: PASS ===
=== ALL PASS ===
```
