+++ .specs/draft-auto-scaling-rule-ux/.context/progress.md
@@ -1,0 +1,20 @@
+# Auto Scaling Rule UX Improvements Progress
+
+## Last Session: 2026-04-13
+
+### 1. Current Phase
+Planning complete. Ready for Wave 1 implementation.
+
+### 2. Next Action
+Begin Sub-task 1: Rename existing editor modal to Legacy, extract Legacy rule list from EndpointDetailPage, add version gating.
+
+### 3. Current Goal
+Begin implementation of auto-scaling rule UX improvements with Strawberry API migration.
+
+### 4. Lessons Learned
+(none yet)
+
+### 5. Completed Work
+- Spec reviewed: .specs/draft-auto-scaling-rule-ux/spec.md
+- Dev plan created: .specs/draft-auto-scaling-rule-ux/dev-plan.md
+- Codebase analyzed: identified all affected files, GraphQL types, version gating patterns
