+++ .specs/draft-auto-scaling-rule-ux/.context/tasks.md
@@ -1,0 +1,18 @@
+# Auto Scaling Rule UX Improvements Tasks
+
+> Last synced: 2026-04-13 (no issue tracker, local dev plan only)
+> Source: Local dev-plan.md
+
+## Progress: 0/4 complete
+
+### Wave 1
+- [ ] Sub-task 1: Rename existing editor modal to Legacy + extract Legacy rule list + version gating
+
+### Wave 2
+- [ ] Sub-task 2: Create new Strawberry AutoScalingRuleList component -- blocked by Sub-task 1
+
+### Wave 3
+- [ ] Sub-task 3: Create new Strawberry AutoScalingRuleEditorModal with single/range mode and Prometheus preset -- blocked by Sub-task 1, Sub-task 2
+
+### Wave 4 (optional)
+- [ ] Sub-task 4: Add Prometheus query result preview in editor modal -- blocked by Sub-task 3
