+++ src/lib/backend.ai-client-esm.ts
@@ -900,6 +900,7 @@ class Client {
       this._features['route-health-status'] = true;
       this._features['model-card-v2'] = true;
       this._features['my-roles'] = true;
+      this._features['prometheus-auto-scaling-rule'] = true;
     }
   }
